+++ tests/conftest.py
@@ -5,3 +5,4 @@ TEST_DIR = Path(__file__).parent
 SAMPLE_DATA_DIR = TEST_DIR / 'sample_data'
 
 from tests.fixtures.model_preds_fixtures import *
+from tests.fixtures.learner_fixtures import *