+++ tonks/__init__.py
@@ -3,5 +3,6 @@ from ._version import __version__
 from tonks.dataloader import MultiDatasetLoader
 from tonks.ensemble import *
 from tonks.learner import MultiTaskLearner, MultiInputMultiTaskLearner
+from tonks.learner_utils import *
 from tonks.text import *
 from tonks.vision import *