+++ tonks/learner_utils/__init__.py
@@ -1,0 +1,1 @@
+from tonks.learner_utils.loss_utils import _get_loss_functions