+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project uses [Semantic Versioning](http://semver.org/).
 
+# [0.4.0] - 2020-4-14
+### Added
+ - Tonks learners can now use multiple loss functions and do multi-label in addition to multi-class tasks
+### Changed
+ - loss functions are specified via a dictionary of tasks and corresponding loss functions 
+
 # [0.3.0] - 2020-4-14
 ### Added
  - read the docs fixes